Sync 8/20/2001 9:27 PDT

+++ lib/mp4v2/mp4common.h
@@ -25,7 +25,7 @@
 #include "mpeg4ip.h"
 
 typedef void*		MP4FileHandle;
-typedef u_int16_t	MP4TrackId;
+typedef u_int32_t	MP4TrackId;
 typedef u_int32_t	MP4SampleId;
 
 #include "mp4util.h"
